apps/backend/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shamchess-backend"
version = "0.1.0"
description = "ShamChess backend service"
authors = [{ name = "ShamChess Team" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "Django==5.0.6",
    "djangorestframework==3.15.1",
    "djangorestframework-simplejwt==5.3.1",
    "channels[daphne]==4.1.0",
    "channels-redis==4.2.0",
    "django-environ==0.11.2",
    "django-cors-headers==4.3.1",
    "psycopg[binary]==3.1.18",
    "redis==5.0.4",
    "Pillow==10.3.0",
    "python-chess==1.999",
    "celery==5.3.6",
    "django-celery-beat==2.5.0",
]

[tool.django]
settings-module = "shamchess_backend.settings"
